+++ src/FunctionHelper.ts
@@ -1,10 +1,11 @@
 import * as vscode from 'vscode'
 import * as fs from 'fs'
 import * as path from 'path'
+import { ExtensionModule } from './ExtensionModule'
+export class FunctionHelper extends ExtensionModule {
 
-export class FunctionHelper implements vscode.Disposable {
-
-    public constructor(context: vscode.ExtensionContext, wsConfig: vscode.WorkspaceConfiguration) {
+    public constructor(context: vscode.ExtensionContext) {
+        super(context)
 
         let changeMethodSignature = vscode.commands.registerCommand('geocpptools.changeFunctionSignature', async (fileUri) => {
             const editor = vscode.window.activeTextEditor!
@@ -14,12 +15,12 @@ export class FunctionHelper implements vscode.Disposable {
             console.log("END CHAR POS:" + editor.selection.end.character)
             console.log("CURSOR POS:" + editor.selection.active)
             console.log("CURSOR POS:" + editor.selection)
-            
+
             let selContent = editor.document.getText(new vscode.Range(editor.selection.start, editor.selection.end))
             console.log("CONTENT" + selContent)
-            
+
             let content = editor.document.getText()
-            
+
             if (editor.selection.isEmpty) {
                 //TODO: Determine method signature
                 let options = {
@@ -32,8 +33,4 @@ export class FunctionHelper implements vscode.Disposable {
         context.subscriptions.push(changeMethodSignature)
     }
 
-    dispose() {
-        //TODO: Implement if necessary
-        console.log('Method not implemented.')
-    }
 }